update migartion and default cutoff time for online filter
